Change button to switch and move to overview
